pyproject.toml
[tool.poetry]
name = "tkcbc"
version = "0.0.0"
description = "TKCBC: Keep Collections Before Cleaning"
authors = ["Kiai <k.kubokawa@klis.tsukuba.ac.jp>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.0"
requests-oauthlib = "^1.3.1"
tqdm = "^4.64.0"
python-dotenv = "^0.20.0"
commentjson = "^0.9.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
bandit = "^1.7.4"
types-requests = "^2.27.31"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
